Initial setup for git info
Git-related information can be now embedded into the source code. CMake
main script first evaluates that possibility (it won-t be possible on a
package redistribution) and then (if that is in fact possible) it will
proceed to pin some useful git information into the config header
(refspec and commit hash). It still needs further evaluation, but so far
so good.