Getting rid of qal functions
From now on, there's no such thing as QAL as we used to know it.
htq2's QAL implementation are just defined symbols of equivalent
OpenAL functions, so that means htq2 uses OpenAL directly and NOT
through function pointers. This implementation has been based on
ioquake3's QAL in which is optional.