Force redefine Windows version to XP
According to msdn, some functions are only available to specific windows
versions and therefore one must explicitly tell the compiler that the
program is being compiled for that particular version. As it turns out,
this is the case of the getaddrinfo() function.
By taking a quick look at the tcp/ip specific extensions for Winsock2
header file (ws2tcpip.h) it's easy to see this OS version dependency .